requirements.txt
chart-studio==1.1.0
plotly==4.6.0
python-dotenv==0.13.0
requests==2.23.0